requirements.txt
# ── Core ──
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
python-dotenv>=1.0.0
pydantic[email]>=2.10.0
pydantic-settings>=2.7.0

# ── Database ──
sqlalchemy>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.0
redis>=5.2.0

# ── Auth ──
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.18

# ── ML / Computer Vision ──
torch>=2.6.0
torchvision>=0.21.0
opencv-python-headless>=4.10.0
numpy>=2.1.0
Pillow>=11.0.0
onnxruntime>=1.20.0

# ── Utilities ──
websockets>=14.1
httpx>=0.28.0

# ── Testing ──
pytest>=8.3.0
pytest-asyncio>=0.24.0
aiosqlite>=0.20.0